examples/ecommerce-search/demo-app/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.0.0
